[UPDATE]: reorganize components for Home page
