+++ client_mobile/package.json
@@ -21,6 +21,7 @@
     "react-native": "0.70.5",
     "react-native-draglist": "^3.0.1",
     "react-native-gesture-handler": "~2.8.0",
+    "react-native-linear-gradient": "^2.6.2",
     "react-native-reanimated": "~2.12.0",
     "react-native-safe-area-context": "4.4.1",
     "react-native-screens": "~3.18.0",
